Loosen validation for array-descendant patch fields
This change allows a PARTIAL_UPDATE or BATCH_PARTIAL_UPDATE request to
set a ReadOnly/CreateOnly field if that field is the descendant of an
array. This is needed because there's no real way to "patch" individual
array elements in Rest.li, so array patching actually sets the entire
array. Thus, users need a way to "patch" an array while maintaining the
existing values of ReadOnly/CreateOnly fields.